+++ changelog.md
@@ -15,6 +15,7 @@ Upcoming release (2022)
     - Confidence intervals added for regression parameters
     - Population plot added with confidence band of regression line
     - :warning: Henze-Zirkler test for assumption of multivariate normality
+    - :warning: White's test and Koenker's test for assumption of homoscedasticity
 - Display the filtered cases when filtering outliers
 - Post hoc Durbin-Conover test after significant Friedman test